refactor/EV-780/Adding folder views to wallet
