+++ lib/main.dart
@@ -1,7 +1,8 @@
-import 'package:crypto/wallet/wallet_page.dart';
 import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+import 'wallet/views/wallet_page.dart';
 
 void main() {
   runApp(
